requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
